Bug fixes, refactor and remove unused files/folders
