+++ index.html
@@ -6,7 +6,6 @@
   <link rel="icon" href="/favicon.ico">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Vite App</title>
-  <link rel="stylesheet" href="./src/style.css">
 </head>
 
 <body>
