+++ src/assets/main.css
@@ -3,7 +3,7 @@
 #app {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
+  padding: 0;
 
   font-weight: normal;
 }
@@ -32,4 +32,47 @@ a,
     grid-template-columns: 1fr 1fr;
     padding: 0 2rem;
   }
+}
+
+* {
+  background-color: darkslategrey;
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+  justify-content: center;
+  text-align: center;
+  padding: 0px;
+  margin: 0px;
+}
+
+.menu {
+  grid-area: menu;
+}
+
+.header {
+  grid-area: header;
+  color: yellow;
+  font-size: 36px;
+  width: 100vw;
+}
+
+p {
+  color: lightyellow;
+}
+
+.main {
+  grid-area: main;
+}
+
+ul {
+  list-style-type: none;
+}
+
+.full-page-layout {
+  display: grid;
+  width: 100vw;
+  height: 100vh;
+  grid-template-areas:
+      'header header header header header header'
+      'menu main main main main main'
+      'footer footer footer footer footer footer';
+  gap: 0;
 }
